instagram_scraper_v2/requirements.txt
playwright>=1.40.0
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
